Unit test created
