+++ src/frontend/src/app/records/alumni/edit-form/edit-form.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, Output, Input, Injectable } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { EventEmitter } from '@angular/core';
-import { AlumniService, AlumniValidator } from '../alumni.service';
-
+import { AlumniService } from '../alumni.service';
+import { AlumniValidator } from '../alumni.validator';
 @Component({
   selector: 'AlumniEditFormComponent',
   templateUrl: './edit-form.component.html',
@@ -22,7 +22,7 @@ formControls = {
     mailing_address_zipcode: new FormControl(''),
     phone_number: new FormControl(''),
     email_address: new FormControl(''),
-    added_by: new FormControl('')
+
 }
 
 
@@ -55,25 +55,47 @@ public subDetailVisible: boolean = false;
     this.currentDetailTab = newTab;
   }
 
+  processSaveResponse(response) {
+
+    if (response.validationErrors) {
+
+
+      this.validationErrors = response.data;
+
+
+    } else if (response.otherErrors) {
+
+
+
+    }
+    else {
+      
+      this.currentStudent.alumni = response.data;
+      this.alumniForm.patchValue(this.currentStudent.alumni);
+    }
+
+
+
+  }
+
+
   submitAlumniGeneralData() {
     
     if (!this.alumniForm.pristine) {
-      console.log(this.alumniForm.value);
+      
       let validator : AlumniValidator = new AlumniValidator();
       this.validationErrors = null;
      let [errorsExist, errors] = validator.validateAlumniRecord(this.alumniForm.value);
      if (errorsExist) {
        this.validationErrors = errors;
      }
-     this.alumniForm.value['updated_by'] = 'TEST';
-     this.alumniForm.value['updated_datetime'] = new Date();
-     
-     this.currentStudent.alumni = this.alumniForm.value;
+     else {
+     this.service.updateAlumni(this.currentStudent.alumni.student_id, this.alumniForm.value).then(result => this.processSaveResponse(result));
      
 
 
     }
     
   }
-
+  }
 }
