+++ src/frontend/src/app/unit-tests/alumni-validation.spec.ts
@@ -1,0 +1,17 @@
+import { AlumniValidator } from '../records/alumni/alumni.validator';
+
+describe('alumni data validation', () => {
+    it('should return validation error', () => {
+        let validator = new AlumniValidator();
+        expect(0).toBe(0);
+        
+
+
+
+
+    })
+
+
+
+
+})
