+++ src/frontend/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compileOnSave": false,
   "compilerOptions": {
-    "baseUrl": "./",
+    "baseUrl": "src",
     "outDir": "./dist/out-tsc",
     "sourceMap": true,
     "declaration": false,
@@ -17,7 +17,11 @@
     "lib": [
       "es2018",
       "dom"
-    ]
+    ],
+   "paths": {
+    "@app/*": ["app/*"]
+   }
+   
   },
   "angularCompilerOptions": {
     "fullTemplateTypeCheck": true,
